+++ packages/core/test/service/fixtures/invalid_case/config.default.ts
@@ -1,0 +1,8 @@
+export const keys = 'demo'
+
+export default {
+  parent: {
+    a: 1,
+    b: 2
+  }
+};