+++ packages/oss/src/index.ts
@@ -1,2 +1,3 @@
 export { OSSConfiguration as Configuration } from './configuration';
 export * from './manager';
+export * from './interface';